requirements.txt
streamlit==1.44.1
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
youtube-transcript-api==0.6.2
pypdf==5.0.0
numpy==1.26.4
python-dotenv==1.0.0
